[#7] Refactor code base
